requirements.txt
streamlit
openai
langchain
langchain-openai
faiss-cpu
python-dotenv
pypdf
docx2txt
tiktoken
